Adding `visualize-watch`: it makes use of `chokidar` to watch for changes and run visualize-all
